_requirements.txt
llama-index-core       # core APIs used: VectorStoreIndex, Settings, StorageContext, query engines
llama-index-vector-stores-chroma  # ChromaVectorStore integration
llama-index-llms-google-genai     # GoogleGenAI LLM driver
llama-index-embeddings-google-genai  # GoogleGenAIEmbedding for text-embedding-004
llama-index-readers-file
llama_index.llms.ollama
chromadb                  # PersistentClient for vector DB
python-dotenv             # load_dotenv for env config
gradio
openpyxl
colorlog